requirements.txt
watchdog==3.0.0
fastapi==0.104.1
uvicorn==0.23.2
sqlalchemy==2.0.22
python-dotenv==1.0.0
